// src/types/contact.ts
// Interface automatically generated by schemas-to-ts

export enum Type {
  Url = 'URL',
  Email = 'Email',
  Phone = 'Phone',
}

export interface Contact {
  id: number;
  attributes: {
    createdAt: Date;
    updatedAt: Date;
    publishedAt?: Date;
    Title?: string;
    Address?: string;
    Type?: Type;
  };
}
export interface Contact_Plain {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Title?: string;
  Address?: string;
  Type?: Type;
}

export interface Contact_NoRelations {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Title?: string;
  Address?: string;
  Type?: Type;
}

export interface Contact_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;
  updatedAt: Date;
  publishedAt?: Date;
  Title?: string;
  Address?: string;
  Type?: Type;
}